requirements.txt
gcloud>=0.18.3
oauth2client>=4.1.3
pycryptodome>=3.15.0
python_jwt>=3.3.2
requests>=2.28.0
requests_toolbelt>=0.9.1
six>=1.16.0
